running exercise with ChoreForm locally and ChoreChart via props
